refactor: rename createScale
